membuat data class dan fungsi untuk menyimpan input data barang

+++ app/src/main/java/com/example/ucproomdb/data/entity/Barang.kt
@@ -10,7 +10,7 @@ data class Barang(
     val idBarang :String,
     val nama :String,
     val deskripsi :String,
-    val harga :Double,
+    val harga :Int,
     val stok :Int,
     val namaSupplier :String,
 )
